+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   to define standard conversations in order to quickly build predefined
   transactions.
 
+### Changed
+* The first occurrence of `:` is used to split an answer into a payee and a
+  narration (eg `Payee: narration`). When no colon is found, all the string is
+  treated as a narration.
+
 ## 0.1.0
 ### Added